requirements.test.txt
python-dateutil==2.9.0.post0
pre-commit
pycares==4.11.0
pytest
pytest-cov
pytest-homeassistant-custom-component
pytest-asyncio
pymodbus==3.11.2
ruff==0.11.5
